# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "家庭记账系统API文档",
    "description": "家庭记账系统后端接口文档",
    "version": "1.0.0",
    "contact": {
      "name": "Family Bill",
      "email": ""
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8090",
      "description": "本地开发环境"
    }
  ],
  "tags": [
    {
      "name": "账单管理",
      "description": "账单相关接口"
    },
    {
      "name": "账单分类管理",
      "description": "账单分类相关接口"
    },
    {
      "name": "支付方式管理",
      "description": "支付方式相关接口"
    }
  ],
  "paths": {
    "/api/bills": {
      "get": {
        "tags": ["账单管理"],
        "summary": "分页查询账单列表",
        "operationId": "getBillPage",
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "description": "开始时间",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "结束时间",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "paymentMethodId",
            "in": "query",
            "description": "支付方式ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "分类ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "类型：1-支出，2-收入",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageNum",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "每页大小",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["账单管理"],
        "summary": "新增账单",
        "operationId": "createBill",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                }
              }
            }
          }
        }
      }
    },
    "/api/bills/{id}": {
      "get": {
        "tags": ["账单管理"],
        "summary": "根据ID查询账单",
        "operationId": "getBillById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "账单ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBillVO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["账单管理"],
        "summary": "更新账单",
        "operationId": "updateBill",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "账单ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["账单管理"],
        "summary": "删除账单",
        "operationId": "deleteBill",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "账单ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/bills/statistics": {
      "get": {
        "tags": ["账单管理"],
        "summary": "获取统计信息",
        "operationId": "getStatistics",
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "description": "开始时间",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "结束时间",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "paymentMethodId",
            "in": "query",
            "description": "支付方式ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "分类ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "类型：1-支出，2-收入",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultStatisticsVO"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": ["账单分类管理"],
        "summary": "查询所有分类",
        "operationId": "getAllCategories",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "类型：1-支出，2-收入",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBillCategoryList"
                }
              }
            }
          }
        }
      }
    },
    "/api/payment-methods": {
      "get": {
        "tags": ["支付方式管理"],
        "summary": "查询所有支付方式",
        "operationId": "getAllPaymentMethods",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "类型：1-支出方式，2-收入方式，3-通用",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPaymentMethodList"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码"
          },
          "message": {
            "type": "string",
            "description": "响应消息"
          },
          "data": {
            "description": "响应数据"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "时间戳"
          }
        }
      },
      "ResultBillVO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/BillVO"
              }
            }
          }
        ]
      },
      "ResultLong": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ]
      },
      "ResultVoid": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "nullable": true
              }
            }
          }
        ]
      },
      "ResultStatisticsVO": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/StatisticsVO"
              }
            }
          }
        ]
      },
      "ResultBillCategoryList": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/BillCategory"
                }
              }
            }
          }
        ]
      },
      "ResultPaymentMethodList": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/PaymentMethod"
                }
              }
            }
          }
        ]
      },
      "PageResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/IPageBillVO"
              }
            }
          }
        ]
      },
      "IPageBillVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BillVO"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "总记录数"
          },
          "size": {
            "type": "integer",
            "description": "每页大小"
          },
          "current": {
            "type": "integer",
            "description": "当前页码"
          },
          "pages": {
            "type": "integer",
            "description": "总页数"
          }
        }
      },
      "BillDTO": {
        "type": "object",
        "required": ["payee", "amount", "transactionTime", "paymentMethodId"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "账单ID"
          },
          "payee": {
            "type": "string",
            "description": "收款方",
            "example": "星巴克咖啡"
          },
          "amount": {
            "type": "number",
            "format": "decimal",
            "description": "金额（支出为负数，收入为正数）",
            "example": -35.50
          },
          "transactionTime": {
            "type": "string",
            "format": "date-time",
            "description": "交易时间"
          },
          "paymentMethodId": {
            "type": "integer",
            "format": "int64",
            "description": "支付方式ID"
          },
          "transactionNo": {
            "type": "string",
            "description": "交易号",
            "example": "WX202412200830001"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "description": "分类ID"
          },
          "remark": {
            "type": "string",
            "description": "备注",
            "example": "早餐咖啡"
          }
        }
      },
      "BillVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "账单ID"
          },
          "payee": {
            "type": "string",
            "description": "收款方"
          },
          "amount": {
            "type": "number",
            "format": "decimal",
            "description": "金额"
          },
          "transactionTime": {
            "type": "string",
            "format": "date-time",
            "description": "交易时间"
          },
          "paymentMethodId": {
            "type": "integer",
            "format": "int64",
            "description": "支付方式ID"
          },
          "paymentMethodName": {
            "type": "string",
            "description": "支付方式名称"
          },
          "transactionNo": {
            "type": "string",
            "description": "交易号"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "description": "分类ID"
          },
          "categoryName": {
            "type": "string",
            "description": "分类名称"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          }
        }
      },
      "StatisticsVO": {
        "type": "object",
        "properties": {
          "totalIncome": {
            "type": "number",
            "format": "decimal",
            "description": "总收入"
          },
          "totalExpense": {
            "type": "number",
            "format": "decimal",
            "description": "总支出"
          },
          "netIncome": {
            "type": "number",
            "format": "decimal",
            "description": "净收入（收入-支出）"
          },
          "categoryStatistics": {
            "type": "array",
            "description": "按分类统计",
            "items": {
              "$ref": "#/components/schemas/CategoryStatistics"
            }
          },
          "paymentMethodStatistics": {
            "type": "array",
            "description": "按支付方式统计",
            "items": {
              "$ref": "#/components/schemas/PaymentMethodStatistics"
            }
          }
        }
      },
      "CategoryStatistics": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "description": "分类ID"
          },
          "categoryName": {
            "type": "string",
            "description": "分类名称"
          },
          "amount": {
            "type": "number",
            "format": "decimal",
            "description": "金额"
          },
          "type": {
            "type": "integer",
            "description": "类型：1-支出，2-收入"
          }
        }
      },
      "PaymentMethodStatistics": {
        "type": "object",
        "properties": {
          "paymentMethodId": {
            "type": "integer",
            "format": "int64",
            "description": "支付方式ID"
          },
          "paymentMethodName": {
            "type": "string",
            "description": "支付方式名称"
          },
          "amount": {
            "type": "number",
            "format": "decimal",
            "description": "金额"
          }
        }
      },
      "BillCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "主键ID"
          },
          "name": {
            "type": "string",
            "description": "分类名称"
          },
          "icon": {
            "type": "string",
            "description": "图标URL"
          },
          "type": {
            "type": "integer",
            "description": "类型：1-支出，2-收入"
          },
          "parentId": {
            "type": "integer",
            "format": "int64",
            "description": "父分类ID，0表示顶级分类"
          },
          "sortOrder": {
            "type": "integer",
            "description": "排序顺序"
          },
          "status": {
            "type": "integer",
            "description": "状态：1-启用，0-禁用"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
      "PaymentMethod": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "主键ID"
          },
          "name": {
            "type": "string",
            "description": "支付方式名称"
          },
          "icon": {
            "type": "string",
            "description": "图标URL"
          },
          "type": {
            "type": "integer",
            "description": "类型：1-支出方式，2-收入方式，3-通用"
          },
          "sortOrder": {
            "type": "integer",
            "description": "排序顺序"
          },
          "status": {
            "type": "integer",
            "description": "状态：1-启用，0-禁用"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      }
    }
  }
}